# cmd/controller/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Qryptic Controller Service to manage the users, groups, gateways and clients.",
        "title": "Qryptic Controller API",
        "contact": {
            "name": "Leetsecure Support",
            "url": "leetsecure.com",
            "email": "hello@leetsecure.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/config": {
            "get": {
                "description": "Get Admin Configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-config"
                ],
                "summary": "GetAdminConfiguration",
                "operationId": "GetAdminConfiguration",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "string collections",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/config/password-login": {
            "put": {
                "description": "update whether password login should be allowed or not",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-config"
                ],
                "summary": "UpdateAllowPasswordLogin",
                "operationId": "UpdateAllowPasswordLogin",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "true or false",
                        "name": "UpdateAllowPasswordLoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.UpdateAllowPasswordLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/config/sso": {
            "post": {
                "description": "Add SSO Configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-config"
                ],
                "summary": "Add SSO Config",
                "operationId": "AddSsoConfig",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "SSO Config Details",
                        "name": "AddSsoConfigRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.AddSsoConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/config/sso-login": {
            "put": {
                "description": "update whether sso login should be allowed or not",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-config"
                ],
                "summary": "UpdateAllowSSOLogin",
                "operationId": "UpdateAllowSSOLogin",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "true or false",
                        "name": "UpdateAllowSSOLoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_leetsecure_kavach-vpn_controller-backend_internal_models.UpdateAllowSSOLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/config/sso/{id}": {
            "delete": {
                "description": "delete sso configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-config"
                ],
                "summary": "DeleteSsoConfig",
                "operationId": "DeleteSsoConfig",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Insert your token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sso id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sso-config": {
            "get": {
                "description": "get-client-ids-of-sso-allowed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "SSO Configs",
                "operationId": "sso-configs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.AddSsoConfigRequest": {
            "type": "object",
            "required": [
                "clientID",
                "clientSecret",
                "domain",
                "platform",
                "provider"
            ],
            "properties": {
                "clientID": {
                    "type": "string"
                },
                "clientSecret": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.UpdateAllowPasswordLoginRequest": {
            "type": "object",
            "required": [
                "allowPasswordLogin"
            ],
            "properties": {
                "allowPasswordLogin": {
                    "type": "boolean"
                }
            }
        },
        "github_com_leetsecure_kavach-vpn_controller-backend_internal_models.UpdateAllowSSOLoginRequest": {
            "type": "object",
            "required": [
                "allowSsoLogin"
            ],
            "properties": {
                "allowSsoLogin": {
                    "type": "boolean"
                }
            }
        },
...
    }
}